Mv source from index to fixWidth.ts and factory.ts
Minor fixes
